03_db/users_service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose
pydantic
email-validator
python-multipart
bcrypt==4.0.1
passlib==1.7.4
